pyproject.toml
[project]
name = "a2-14-oct-25"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiohttp>=3.13.0",
    "beautifulsoup4>=4.14.2",
    "html5lib>=1.1",
    "ipykernel>=7.0.1",
    "jupyterlab>=4.4.10",
    "lxml>=6.0.2",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "urllib3==2.5.0",
]
